plugin(locale): get the currency symbol for the current locale
There is no clean way to retrieve solely the currency symbol. Therefore a pivot element was chosen to
create a formatted number. In that result the pivot element was removed to retrieve only the currency
symbol.
Depending on the locale the symbol can be on either the left or the right side of the number.

Fixes #339