feat(transloco): 🎸 Add config option for scope auto prefixing (#868)
Add an option to the config, to control if the TranslateService should
prefix (or not) the translation key with the scope when provided. It fix
an issue when a scope has been provided, the root translations keys
can't be used with the TranslocoService methods due to auto-prefixing of
the scope to them.

✅ Closes: #855