fix(transloco): 🐛 Avoid toObservable on signal API + string key
Simplifies the common use case of static keys / params for the signal
APIs by avoiding the toObservable call, which uses the angular effect
API to listen to changes from the signal and propogate those into the
observable stream. Using the effect call causes the component to render
twice, once with an empty string as the signal value and the second time
asyncronously with the actual value.

✅ Closes: #894